Refactor : journeyDetail.js -> JourneyDetailContext.jsx 이름 변경
